# doc/api.yml
openapi: 3.0.3
info:
  title: SchoolM API
  description: School management API (HEI)
  version: v1
security:
  - bearerAuth: []
paths:
...
  /courses:
    get:
      tags:
        - Course
      summary: List all courses
      description: Returns the full list of courses.
      operationId: getAllCourses
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
    put:
      tags:
        - Course
      summary: Create or update a course
      description: >-
        Creates a new course or updates an existing one. If the request body
        contains an id, the existing course is updated. Otherwise, a new
        course is created.
      operationId: upsertCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseRequest'
      responses:
        '200':
          description: Course created or updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
  /courses/{courseId}:
    get:
      tags:
        - Course
      summary: Get a course by id
      description: Returns a single course by its identifier.
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          required: true
          description: Course identifier
          schema:
            type: string
      responses:
        '200':
          description: Course retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Course
      summary: Delete a course
      description: >-
        Deletes a course along with its exams and their grades and grade
        history.
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          required: true
          description: Course identifier
          schema:
            type: string
      responses:
        '204':
          description: Course deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/teacher:
    put:
      tags:
        - Course
      summary: Assign teachers to a course
      description: Replaces the full set of teachers assigned to the course.
      operationId: assignTeachers
      parameters:
        - name: courseId
          in: path
          required: true
          description: Course identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherIdsRequest'
      responses:
        '200':
          description: Teachers assigned to the course successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/group:
    put:
      tags:
        - Course
      summary: Assign groups to a course
      description: Replaces the full set of groups that take the course.
      operationId: assignGroups
      parameters:
        - name: courseId
          in: path
          required: true
          description: Course identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupIdsRequest'
      responses:
        '200':
          description: Groups assigned to the course successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/exams:
    get:
      tags:
        - Exam
      summary: List exams of a course
      description: Returns the exams of the given course.
      operationId: getCourseExams
      parameters:
        - name: courseId
          in: path
          required: true
          description: Course identifier
          schema:
            type: string
      responses:
        '200':
          description: Exams retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exam'
        '404':
          $ref: '#/components/responses/NotFound'
  /exams/{id}:
    get:
      tags:
        - Exam
      summary: Get an exam by id
      description: Returns a single exam by its identifier.
      operationId: getExamById
      parameters:
        - name: id
          in: path
          required: true
          description: Exam identifier
          schema:
            type: string
      responses:
        '200':
          description: Exam retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Exam
      summary: Delete an exam
      description: >-
        Deletes an exam along with its grades and grade history.
      operationId: deleteExam
      parameters:
        - name: id
          in: path
          required: true
          description: Exam identifier
          schema:
            type: string
      responses:
        '204':
          description: Exam deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /exams:
    put:
      tags:
        - Exam
      summary: Create or update an exam
      description: >-
        Creates a new exam or updates an existing one. If the request body
        contains an id, the existing exam is updated.
      operationId: upsertExam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamRequest'
      responses:
        '200':
          description: Exam created or updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TeacherIdsRequest:
      type: object
      required:
        - teacherIds
      properties:
        teacherIds:
          type: array
          minItems: 1
          items:
            type: string
          description: List of teacher identifiers to assign to the course
    GroupIdsRequest:
      type: object
      required:
        - groupIds
      properties:
        groupIds:
          type: array
          minItems: 1
          items:
            type: string
          description: List of group identifiers that take the course
...
    Course:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
        title:
          type: string
        credit:
          type: integer
        track:
          type: string
        semester:
          type: string
        teacherIds:
          type: array
          items:
            type: string
          description: Teachers assigned to the course
        groupIds:
          type: array
          items:
            type: string
          description: Groups that take the course
    CourseRequest:
      type: object
      required:
        - ref
        - title
        - credit
        - track
        - semester
      properties:
        id:
          type: string
          description: Identifier, present when updating an existing course
        ref:
          type: string
        title:
          type: string
        credit:
          type: integer
        track:
          type: string
          description: Track of the course (COMMON, EL or TN)
        semester:
          type: string
          description: Semester of the course (S1 to S6)
...
    Exam:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        dateExam:
          type: string
          format: date
        coefNumerator:
          type: integer
        coefDenominator:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExamRequest:
      type: object
      required:
        - courseId
        - dateExam
        - coefNumerator
        - coefDenominator
      properties:
        id:
          type: string
          description: Identifier, present when updating an existing exam
        courseId:
          type: string
          description: Identifier of the course the exam belongs to
        dateExam:
          type: string
          format: date
        coefNumerator:
          type: integer
        coefDenominator:
          type: integer
...
    ErrorBody:
      type: object
      properties:
        status:
          type: integer
        reason:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'